requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
